+++ src/WinBridge.Runtime.Windows.Display/DisplayIdentityFailureSemantics.cs
@@ -5,7 +5,8 @@ namespace WinBridge.Runtime.Windows.Display;
 internal static class DisplayIdentityFailureSemantics
 {
     public static bool IsIdentityBreaking(string? failedStage) =>
-        failedStage is DisplayIdentityFailureStageValues.GetBufferSizes
+        failedStage is DisplayIdentityFailureStageValues.GetMonitorInfo
+            or DisplayIdentityFailureStageValues.GetBufferSizes
             or DisplayIdentityFailureStageValues.QueryDisplayConfig
             or DisplayIdentityFailureStageValues.GetSourceName
             or DisplayIdentityFailureStageValues.CoverageGap;
@@ -13,8 +14,9 @@ internal static class DisplayIdentityFailureSemantics
     public static int GetPriority(string? failedStage) =>
         failedStage switch
         {
-            DisplayIdentityFailureStageValues.GetBufferSizes => 400,
-            DisplayIdentityFailureStageValues.QueryDisplayConfig => 300,
+            DisplayIdentityFailureStageValues.GetBufferSizes => 500,
+            DisplayIdentityFailureStageValues.QueryDisplayConfig => 400,
+            DisplayIdentityFailureStageValues.GetMonitorInfo => 300,
             DisplayIdentityFailureStageValues.GetSourceName => 200,
             DisplayIdentityFailureStageValues.CoverageGap => 100,
             DisplayIdentityFailureStageValues.GetTargetName => 50,